server/requirements.txt
fastapi
uvicorn
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
firebase-admin
aiosmtplib
redis
PyJWT[crypto]
pydantic[email]
h3